==> Project/1_rating_word_offensiveness.ipynb <==
## Setting up a Dataframe for the rating of offensiveness
I'm using a paper released by the British telecommunications regulator Ofcom: https://www.ofcom.org.uk/__data/assets/pdf_file/0023/91625/OfcomQRG-AOC.pdf

There is a useful transcription here:
http://metro.co.uk/2016/10/02/swearing-ranked-from-mild-to-strongest-6165629/#

I created a JSON file from the data of that website.
This notebook will load this file and convert it to a Dataframe with a usable layout.

import json

with open("../datasets/metro_co_uk_transcription.json") as file:
    offensiveness_rating = json.load(file)
    
# the offensiveness rating is a nested JSON file
# after three levels of json objects, there is a list of words
# it looks like this:
print(offensiveness_rating["discriminatory"]["religion"]["strong"])
print(offensiveness_rating["discriminatory"]["race"]["medium"])

# there are also offensive words that are not discriminating
# for them, the second key doesn't really make sense, I just entered "offensive"
print(offensiveness_rating["non-discriminatory"]["offensive"]["mild"])

# for further processing, we want to have a dataframe with the words as index
# a cell in the dataframe should be a word and the rating of offensiveness
# if the word is discriminating, we also want to have the target group

entries = []

def clean(word):
    return word.strip().lower()

# words are always three levels down in the file
for category, value in offensiveness_rating.items():
    for target, value in value.items():
        for strength, words in value.items():
            for word in words:
                entry = {}
                entry["category"] = clean(category)
                entry["word"] = clean(word)
                entry["strength"] = clean(strength)
                
                assert category in ["discriminatory", "non-discriminatory"]
                
                if category=="discriminatory":
                    entry["target"] = clean(target)
                else:
                    entry["target"] = None
                
                entries.append(entry)

entries[:2]

import pandas as pd

word_table = pd.DataFrame(entries)
word_table.head(5)

word_table= word_table.set_index("word")
word_table.head(5)

word_table.to_pickle("../pickles/word_table.pickle")
