==> Project/4_restructuring_word_rating.ipynb <==
## Restructure our table

We flatten the table of offensive words. This means that we can now use it in combination with technologies such as spark - that don't support multiindex structures.

import pandas as pd

word_table = pd.read_pickle("../pickles/word_table_cleaned.pickle")

word_table.head()

word_table["descriptor"] = word_table[["category", "strength", "target"]].apply(tuple, axis=1)

word_table.head()

word_table_filtered = word_table["descriptor"]

word_table_filtered.head()

word_table_filtered.to_csv("../datasets/word_table.csv", header=True)
